requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_jwt_extended
flask-cors
psycopg2-binary
python-dotenv
python-dateutil==2.9.0.post0